Basic architecture
Created an Enemy class that has a state machine. The state machine is a simple one right now, built on enums.

There are two ways of configuring enemies: the serializable fields `health` and `damage` and an EnemyConfig scriptable object. I will need to choose one way to configure over another later.

Enemies will be created by an EnemyFactory. It's a MonoBehavior for now, but that's not great, I'll later initialize it with Zenject.

All enemies are controlled by the EnemySystem, it's useful for controlling all the enemies in the game at once.