Fix: Added collider to door arch
Door arch had a mesh collider before that, but it didn't work for some reason. Because of that, the player could walk through walls and fall from the map completely.

I replaced the mesh collider with a simple box collider. Just to keep in mind: because I didn't add a collider to the top of the arch, the player can jump through it.