Fix: Bullets not detecting collisions
This was because the bullet's rigidbodies were in kinematic mode. Now they hit the terrain and the enemy.

When fixing, I also encountered that when cacodemon shoots, the bullet collides with the cacodemon himself: a common bug.  I couldn't reuse the bullet layer because it ignores collisions with the player, so I created a brand new layer for enemy bullets and set it up to ignore player's bullets and the enemy.