refactor: Make pool generic
Now Pool will store IPoolable objects instead of a IBullet. An abstract class Poolable<T> implements all the duplicate logic for IPoolable. Now, if you need to store some object in a pool, you can inherit from Poolable<T> with T being the type you need, implement GetObject like this:

`GetObject() => this;`

and a ReleaseRequested event that can be invoked later to release the object from the pool.

Also simplified the code that deals with releasing objects from the pool: before that, bullet managed it's release state itself. I remade the pool so there was no longer an error with releasing an object that was already released and now this logic in Bullet is no longer needed.