fix: Animations stop working in some cases
This was caused by the faulty MeleeHit animation logic. I Resetted the state when the hit was dealt and so the player could melee hit twice, which freezed all the animation.

This fix however now causes the melee hit animation to be played completely before giving control back to the player. It feels a lot worse, but solving this would require creating one more flag, which is also not good. Again, state machine/behavior tree will work a lot better in this situation.