Refactor/kawaii code/pool (#18)
* refactor: Make pool generic

Now Pool will store IPoolable objects instead of a IBullet. An abstract class Poolable<T> implements all the duplicate logic for IPoolable. Now, if you need to store some object in a pool, you can inherit from Poolable<T> with T being the type you need, implement GetObject like this:

`GetObject() => this;`

and a ReleaseRequested event that can be invoked later to release the object from the pool.

Also simplified the code that deals with releasing objects from the pool: before that, bullet managed it's release state itself. I remade the pool so there was no longer an error with releasing an object that was already released and now this logic in Bullet is no longer needed.

* refactor: Use BulletBuilder instead of BulletFactory

Since Builder was already used everywhere else, there was no point having a separate Factory class. This change should have simplified the code a little.

* perfomance: Set Pool's max size to ammo count

* refactor: Use Factory instead of a Builder

This was more of a naming problem: the builder wasn't being used for any configuration and was being used like a factory.

* fix: Animations stop working in some cases

This was caused by the faulty MeleeHit animation logic. I Resetted the state when the hit was dealt and so the player could melee hit twice, which freezed all the animation.

This fix however now causes the melee hit animation to be played completely before giving control back to the player. It feels a lot worse, but solving this would require creating one more flag, which is also not good. Again, state machine/behavior tree will work a lot better in this situation.