Feature/kawaii code/damage system (#19)
* Add damage system

A simple implementation, based on an IDamageable interface.

* Add some damage feedback

The enemy will enter a small stun when shot and the player's health and armor will now display properly.

* Make collectables give bonuses

Collectabilities was a sore in the eye, so I renamed it

* Melee hit now deals damage

* refactor + fix

Successive melee hits now finally don't freeze the animations. I claimed to fix this before, but I was wrong. The cause of the bug was that I forgot to add a transition from the Hit animation to Idle in the Animator.